Extract out common function for resizeable panels view - Also fix bug around horizontal resizing of the window
